# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Library System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  admin: 'admin',
  librarian: 'librarian',
  instructor: 'instructor',
  student: 'student',
} as const;

export interface RegisterInput {
  fullname: string;
  email: string;
  password: string;
  phone: string;
  address: string;
  campus: string;
  role: RegisterInputRole;
  /** @nullable */
  studentNumber?: string | null;
  /** @nullable */
  course?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  section?: string | null;
}

export interface User {
  id: number;
  fullname: string;
  email: string;
  phone: string;
  address: string;
  campus: string;
  role: string;
  /** @nullable */
  studentNumber?: string | null;
  /** @nullable */
  course?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  section?: string | null;
  isApproved?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  fullname?: string;
  phone?: string;
  address?: string;
  campus?: string;
  isApproved?: boolean;
  /** @nullable */
  course?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  section?: string | null;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  description: string;
  category: string;
  campus: string;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  isbn?: string | null;
  /** @nullable */
  publishedYear?: number | null;
  isAvailablePhysical?: boolean;
  totalCopies?: number;
  availableCopies?: number;
  createdAt: string;
}

export interface BookInput {
  title: string;
  author: string;
  description: string;
  category: string;
  campus: string;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  isbn?: string | null;
  /** @nullable */
  publishedYear?: number | null;
  isAvailablePhysical?: boolean;
  totalCopies?: number;
}

export interface BookUpdate {
  title?: string;
  author?: string;
  description?: string;
  category?: string;
  campus?: string;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  isbn?: string | null;
  /** @nullable */
  publishedYear?: number | null;
  isAvailablePhysical?: boolean;
  totalCopies?: number;
}

export interface MyListItem {
  id: number;
  userId: number;
  bookId: number;
  addedAt: string;
  book: Book;
}

export interface MyListInput {
  bookId: number;
}

export type BorrowRecordStatus = typeof BorrowRecordStatus[keyof typeof BorrowRecordStatus];


export const BorrowRecordStatus = {
  borrowed: 'borrowed',
  returned: 'returned',
  overdue: 'overdue',
} as const;

export interface BorrowRecord {
  id: number;
  userId: number;
  bookId: number;
  borrowedAt: string;
  /** @nullable */
  returnedAt?: string | null;
  dueDate?: string;
  status: BorrowRecordStatus;
  user?: User;
  book?: Book;
}

export interface BorrowInput {
  userId: number;
  bookId: number;
  dueDate: string;
}

export type BorrowUpdateStatus = typeof BorrowUpdateStatus[keyof typeof BorrowUpdateStatus];


export const BorrowUpdateStatus = {
  borrowed: 'borrowed',
  returned: 'returned',
  overdue: 'overdue',
} as const;

export interface BorrowUpdate {
  status?: BorrowUpdateStatus;
  /** @nullable */
  returnedAt?: string | null;
}

export interface MonitoringStats {
  totalUsers: number;
  totalBooks: number;
  activeBorrows: number;
  totalStudents: number;
  totalInstructors: number;
  overdueBooks?: number;
  totalCampuses?: number;
}

export interface CampusStat {
  campus: string;
  totalUsers: number;
  students: number;
  instructors: number;
  activeBorrows: number;
}

export interface CourseStat {
  course: string;
  year: string;
  section: string;
  studentCount: number;
  campus: string;
  activeBorrows?: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  userName?: string;
  /** @nullable */
  bookTitle?: string | null;
  campus?: string;
  createdAt: string;
}

export type ListUsersParams = {
campus?: string;
role?: string;
course?: string;
year?: string;
section?: string;
};

export type ListBooksParams = {
search?: string;
category?: string;
campus?: string;
};

export type ListBorrowRecordsParams = {
userId?: number;
status?: string;
};