# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Library System API
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  title: string;
  author: string;
  description: string;
  category: string;
  campus: string;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  isbn?: string | null;
  /** @nullable */
  publishedYear?: number | null;
  isAvailablePhysical?: boolean;
  totalCopies?: number;
}