# lib/api-zod/src/generated/types/borrowInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Library System API
 * OpenAPI spec version: 0.1.0
 */

export interface BorrowInput {
  userId: number;
  bookId: number;
  dueDate: string;
}