# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Library System API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullname: string;
  email: string;
  phone: string;
  address: string;
  campus: string;
  role: string;
  /** @nullable */
  studentNumber?: string | null;
  /** @nullable */
  course?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  section?: string | null;
  isApproved?: boolean;
  createdAt: string;
}